+++ src/me/anno/objects/meshes/MeshData.kt
@@ -6,6 +6,6 @@ import org.joml.Vector4f
 
 abstract class MeshData: CacheData {
 
-    abstract fun draw(stack: Matrix4fArrayList, time: Float, color: Vector4f)
+    abstract fun draw(stack: Matrix4fArrayList, time: Double, color: Vector4f)
 
 }
